Add files via upload
CS410 Real Estate Calculator project. Test page allows user to select a city or draw a polyline and return all the properties for sale with the area. Each property will pop up with a marker which indicates various prices relative to the real estate broker and the number of bedrooms and bathrooms for the property.